+++ packages/zaml-parser/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.10] - 2019-01-28
+
+### Fixed
+
+* Fix literal parsing
+  * Disable truthy boolean in metadata
+  * Parse big integer as string
+  * Correctly parse string with leading number
+  * Added literal parsing unit tests
+
+### Added
+
+* Command line parser debugger
+
 ## [0.6.9] - 2019-01-25
 
 ### Fixed